requirements.txt
# ─── Core ─────────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0

# ─── Docker ───────────────────────────────────────────────────────────────────
docker>=7.0.0

# ─── LLM Providers ────────────────────────────────────────────────────────────
langchain-core>=0.2.0
langchain-ollama>=0.1.0
langchain-anthropic>=0.1.0
langchain-openai>=0.1.0          # covers openai, groq, deepseek, openrouter, together
langchain-mistralai>=0.1.0
langchain-google-genai>=1.0.0

# ─── Memory ───────────────────────────────────────────────────────────────────
chromadb>=0.5.0

# ─── Security ─────────────────────────────────────────────────────────────────
# (No extra deps — uses stdlib re, hashlib, secrets)
